==> Dominance Data (Study 2)/1.One-zero sampling analysis.Rmd <==
---
title: "*1.One-zero sampling analysis"
author: "Justin Varholick"
date: "2023-10-10"
output: html_document
---

```{r setup}
library(tidyverse)
library(ggplot2)
library(see)
library(cowplot)
library(readxl)

library(FSA)
library(lme4)
library(lmerTest)
library(afex) 
```

# Extract data from BORIS
1 Open project file in BORIS
2 Click the Observations Menu > Export Events > aggregated events
3 Select the observations you wish to export, then all subjects and all behaviors
4 Click YES for group events into one file
5 Save as .csv file

One file for gg and one for jv*
*select all gg, and then only select the necessary obs for jv

```{r data import}
data_master <- read.csv("Data/Ag_Boris_20211109_fromMaster_B1B2B3.csv")

not_reliability <- c("jv2", "gg2")

data_1 <- data_master %>% 
  separate(col=Observation.id, into=c("Cage", "Date", "Hour", "Coder"), sep="_") %>% 
  filter(!Coder %in% not_reliability) %>% 
  select(Cage, Date, Hour, Coder, Behavior, Start..s.)

head(data_1)
levels(as.factor(data_1$Coder))

data_2 <- data_1 %>% 
  filter(Cage!="1") %>% 
  filter(Date!=618) %>% 
  droplevels()

data_2$Cage <- as.factor(data_2$Cage)
data_2$Date <- as.factor(data_2$Date)
data_2$Hour <- as.factor(data_2$Hour)
data_2$Coder <- as.factor(data_2$Coder)
levels(data_2$Cage)
levels(data_2$Date)
levels(data_2$Hour)
levels(data_2$Coder)
data_2 <- mutate(data_2, minute=round(Start..s./60))

sex_id <- read_excel("Data/AcomysDom_Inventory.xlsx", sheet = "Sheet3")
sex_id$Cage <- as.factor(sex_id$Cage)

data_3 <- left_join(data_2, sex_id, by="Cage")


huddle <- c("Side Huddle", "Mounted Huddle")

data_3 <- data_3 %>% 
  mutate(Behavioral.category = case_when(Behavior == "Activity" ~ "Activity",
                                         !Behavior %in% huddle ~ "Agonistic",
                                         Behavior %in% huddle ~ "Huddling") ) 
```
# Day and night plot to justify exclusive night coding
```{r, filter batch1, make graph}

batch1 <- c(957162, 922587, 941702, 927693)
batch1data <- data_3 %>% 
  filter(Cage %in% batch1) %>% 
  filter(Behavior == "Activity") %>% 
  group_by(Cage, Sex, Date, Hour) %>% 
  count()
batch1data <- mutate(batch1data, propact=(n/15)*100)
batch1data <- mutate(batch1data, species="Acomys")
batch1data$Hour <- factor(batch1data$Hour,levels=c("20","21","22","23","00","01","02","03","04","05","06","07","08","09","10","11","12","13","14","15","16","17","18","19"))

heatactivity <- ggplot(batch1data, aes(Hour, Date, fill=propact)) +
  geom_tile(color = "black") +
  geom_vline(xintercept="06", linetype="dashed", color="black")+
  scale_fill_viridis_c() +
  theme_cowplot()+
  labs(fill = "Estimated \ntime active (%)")+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
heatactivity
```

```{r write plot}
ggsave("Figures/DayNightHeatMap.pdf", heatactivity, unit = "mm", height = 180, width = 180, dpi = 300)
ggsave("Figures/DayNightHeatMap.png", heatactivity, unit = "mm", height = 180, width = 180, dpi = 300)
```

# Behavior plots for groups 
```{r}
lightness <- c("06","07","08","09","10","11","12","13","14","15","16","17","18","19")

data_4 <- data_3 %>% 
  filter(!Hour %in% lightness) %>% 
  group_by(Cage, Trt, Date, Hour, Behavior) %>% 
  count() %>% 
  mutate(nadj = case_when(
    n > 15 ~ 15, n <= 15 ~ n)) %>% 
  mutate(prop = (nadj/15)*100) %>% 
  ungroup()

zeromatrix <- data_4 %>% 
  select(Cage, Trt, Date, Hour) %>% 
  mutate(Activity=0) %>% 
  mutate(Chasing=0) %>% 
  mutate(Displace=0) %>% 
  mutate(`Mounting (active)`=0) %>%
  mutate(`Mounted Huddle`=0) %>% 
  mutate(`Side Huddle`=0) %>% 
  gather("Behavior", "n_prime", 5:10) %>% 
  distinct()
  
data_5 <- left_join(zeromatrix, data_4, by=c("Cage", "Trt", "Date", "Hour", "Behavior")) 
data_5[is.na(data_5)]<-0 

data_5 <- data_5 %>% 
  mutate(Trt2 = case_when(
    Trt == "YFP" ~ "Female Pairs",
    Trt == "YMP" ~ "Male Pairs",
    Trt == "YMT" ~ "Male Triads",
    Trt == "AFP" ~ "Aged Females"
  )) %>% 
  mutate(Behavioral.category = case_when(Behavior == "Activity" ~ "Activity",
                                         !Behavior %in% huddle ~ "Agonistic",
                                         Behavior %in% huddle ~ "Huddling") ) 
  

data_5$Hour <- factor(data_5$Hour, levels=c("20","21","22","23","00","01","02","03","04","05","06","07","08","09","10","11","12","13","14","15","16","17","18","19"))

data_5$Behavior <- factor(data_5$Behavior,
                          levels = c("Activity", "Chasing", "Displace", "Mounting (active)", "Side Huddle", "Mounted Huddle"))

levels(data_5$Behavior) <- c("Activity", "Chasing", "Displace", "Mounting \n(active)",
                             "Side \nHuddle", "Mounted \nHuddle")

data_5$Trt2 <- factor(data_5$Trt2,
                     levels = c("Male Pairs", "Female Pairs", "Male Triads", "Aged Females"))

#remove triads
data_5 <- filter(data_5, Trt != "YMT")

write.csv(data_5, "data_5.csv")

```


```{r}

night1 <- c("20190730", "20200121", "20200915")
morn1 <- c("20190731", "20200122", "20200916")
night2 <- c("20190731", "20200122", "20200916")
morn2 <- c("20190801", "20200123", "20200917")
night3 <- c("20190801", "20200123", "20200917")
morn3 <- c("20190802", "20200124", "20200918")

night4 <- c("20190806", "20200130", "20200922")
morn4 <- c("20190807", "20200131", "20200923")
night5 <- c("20190807", "20200131", "20200923")
morn5 <- c("20190808", "20200201", "20200924")
night6 <- c("20190808", "20200201", "20200924")
morn6 <- c("20190809", "20200202", "20200925")
night6.5 <- ("20200202")
morn6.5 <- ("20200203")

night7 <- c("20190815", "20200204", "20200929")
morn7 <- c("20190816", "20200205", "20200930")
night8 <- c("20190816", "20200205", "20200930")
morn8 <- c("20190817", "20200206", "20201001")
night9 <- c("20190817", "20200206", "20201001")
morn9 <- c("20190818", "20200207", "20201002")
night10 <- c("20190818")
morn10 <- c("20190819")

#alter the date to only be an integer
data_51 <- data_5 %>% 
  mutate(Hour2 = case_when(Hour == "00" ~ 0,
         Hour == "01" ~ 1,
         Hour == "02" ~ 2,
         Hour == "03" ~ 3,
         Hour == "04" ~ 4,
         Hour == "05" ~ 5,
         Hour == "20" ~ 20,
         Hour == "21" ~ 21,
         Hour == "22" ~ 22,
         Hour == "23" ~ 23)) %>% 
  mutate(Date2 = case_when((Date %in% night1 & Hour2 > 19) ~ 1,
                           (Date %in% morn1 & Hour2 < 19) ~ 1,
                           (Date %in% night2 & Hour2 > 19) ~ 2,
                           (Date %in% morn2 & Hour2 < 19) ~ 2,
                           (Date %in% night3 & Hour2 > 19) ~ 3,
                           (Date %in% morn3 & Hour2 < 19) ~ 3,
                           (Date %in% night4 & Hour2 > 19) ~ 4,
                           (Date %in% morn4 & Hour2 < 19) ~ 4,
                           (Date %in% night5 & Hour2 > 19) ~ 5,
                           (Date %in% morn5 & Hour2 < 19) ~ 5,
                           (Date %in% night6 & Hour2 > 19) ~ 6,
                           (Date %in% morn6 & Hour2 < 19) ~ 6,
                           (Date %in% night6.5 & Hour2 > 19) ~ 6.5,
                           (Date %in% morn6.5 & Hour2 < 19) ~ 6.5,
                           (Date %in% night7 & Hour2 > 19) ~ 7,
                           (Date %in% morn7 & Hour2 < 19) ~ 7,
                           (Date %in% night8 & Hour2 > 19) ~ 8,
                           (Date %in% morn8 & Hour2 < 19) ~ 8,
                           (Date %in% night9 & Hour2 > 19) ~ 9,
                           (Date %in% morn9 & Hour2 < 19) ~ 9,
                           (Date %in% night10 & Hour2 > 19) ~ 10,
                           (Date %in% morn10 & Hour2 < 19) ~ 10))

```

```{r}
#mean per cage
data_6 <- data_5 %>% 
  group_by(Cage, Trt, Trt2, Behavior) %>% 
  summarise(mean_prop = mean(prop))

groupbhv <- ggplot(data_6, aes(Trt2, mean_prop, fill = Trt2)) +
  stat_summary(geom="bar", position = "identity", fun = "mean", color = "black", alpha = 0.8) +
  stat_summary(geom="errorbar", fun.data = "mean_se", width = 0.2) +
  geom_point(size = 1) +
  scale_y_continuous(limits = c(0,100), breaks = c(seq(0,100, by=10)), expand = expansion(mult = c(0.02,0)))+
  ylab("% behavior \nfirst 15-min / hour") + xlab("") +
  facet_grid(cols=vars(Behavior)) +
  scale_fill_oi() +
  theme_minimal_hgrid() + theme(axis.text.x = element_text(angle = 90, vjust = .5, hjust = 1), legend.position = "none") + panel_border()
groupbhv

#groupbhvbox <- ggplot(data_6, aes(Behavior, mean_prop, fill=Trt2)) +
#  geom_boxplot() +
#  scale_y_continuous(limits = c(0,100))+
#  ylab("Estimated time (%/hour)") +
#  facet_grid(rows = vars(Trt2)) +
#  scale_fill_discrete() +
#  theme_cowplot() +
#  theme(legend.position = "none")
#groupbhvbox

```

```{r}

chases <- data_5 %>% 
  filter(Behavior == "Chasing") %>% 
  group_by(Trt2, Behavior, Hour) %>% 
  summarise(mean_prop = mean(prop))

chases_final <- filter(chases, Hour == "05")
chases_final$mean_prop <- c(5.5,12,-1)

hourplot <- ggplot(chases, aes(Hour, mean_prop, color=Trt2, group = Trt2)) +
  geom_line(size = 1) +
  scale_y_continuous(limits=c(0,100), expand = expansion(mult = c(0.05,0)),
                     name = "% chasing \nfirst 15-min",
                     sec.axis = dup_axis(
                       breaks = chases_final$mean_prop,
                       labels = c("Male Pairs", "Female Pairs", "Aged Females"),
                       name = NULL
                     )) +
  scale_x_discrete(expand = expansion(mult = c(0.02,0)),
                   breaks = c("20", "21", "22", "23", "00", "01", "02", "03", "04", "05"),
                   labels = c("20h", "21h", "22h", "23h", "00h", "01h", "02h", "03h", "04h", "05h"),
                   name = NULL) +
  scale_color_oi()+ scale_fill_oi() +
  theme_cowplot() + xlab("Hour of night") +
  theme(legend.position = "none",
        axis.line.y.right = element_blank(),
        axis.ticks.y.right = element_blank())
hourplot
```
```{r}

panel <- plot_grid(groupbhv, hourplot,
                   rel_heights = c(1.5,1),
          nrow = 2, labels = "AUTO")
panel

ggsave("Figures/GroupBehaviorPlot.pdf", panel, unit = "mm", height = 180, width = 180, dpi = 300)
ggsave("Figures/GroupBehaviorPlot.png", panel, unit = "mm", height = 180, width = 180, dpi = 300)
```
```{r}

data_61 <- data_51 %>% 
  group_by(Cage, Trt2, Behavior, Date2) %>% 
  summarise(mean_prop = mean(prop))
  

ggplot(data_61, aes(Date2, mean_prop, color = Behavior)) +
  geom_point() +
  facet_grid(cols = vars(Cage), rows = vars(Behavior), scale = "free")
```



#stats
```{r}
data_7 <- data_51 %>% 
  group_by(Cage, Trt2, Date2, Behavior) %>% 
  mutate(mean_prop = mean(prop)) %>% 
  select(Trt2, Date2, Behavior, mean_prop) %>% 
  unique()

ggplot(data_7, aes(mean_prop)) + geom_histogram() + facet_wrap(~Behavior)


```
```{r}

#just log transform the data and run the lmer model. It won't working because you are running proportions not the count data it is expecting.

data_transform <- data_7 %>% 
  mutate(mean_prop_c = (mean_prop + 1)) %>% 
  mutate(mean_prop_log = log(mean_prop_c))
ggplot(data_transform, aes(mean_prop_log)) + geom_histogram() + facet_wrap(~Behavior)


act_mdl <- lmer(mean_prop_log ~ Trt2*Date2 + (1|Cage), data = filter(data_transform, Behavior == "Activity"))
anova(act_mdl, test="F")

chs_mdl <- lmer(mean_prop_log ~ Trt2*Date2 + (1|Cage), data = filter(data_transform, Behavior == "Chasing"))
anova(chs_mdl, test="F")

dsp_mdl <- lmer(mean_prop_log ~ Trt2*Date2 + (1|Cage), data = filter(data_transform, Behavior == "Displace"))
anova(dsp_mdl, test="F")

mnt_mdl <- lmer(mean_prop_log ~ Trt2*Date2 + (1|Cage), data = filter(data_transform, Behavior == "Mounting \n(active)"))
anova(mnt_mdl, test="F")

shudd_mdl <- lmer(mean_prop_log ~ Trt2*Date2 + (1|Cage), data = filter(data_transform, Behavior == "Side \nHuddle"))
anova(shudd_mdl, test="F")

mhudd_mdl <- lmer(mean_prop_log ~ Trt2*Date2 + (1|Cage), data = filter(data_transform, Behavior == "Mounted \nHuddle"))
anova(mhudd_mdl, test="F")
summary(mhudd_mdl)

```


```{r}

print("Activity")
kruskal.test(mean_prop ~ Trt2, data = filter(data_6, Behavior == "Activity"))
print("Chase")
kruskal.test(mean_prop ~ Trt2, data = filter(data_6, Behavior == "Chasing"))
print("Displace")
kruskal.test(mean_prop ~ Trt2, data = filter(data_6, Behavior == "Displace"))
print("Mounting")
kruskal.test(mean_prop ~ Trt2, data = filter(data_6, Behavior == "Mounting \n(active)"))
print("SideHuddle")
kruskal.test(mean_prop ~ Trt2, data = filter(data_6, Behavior == "Side \nHuddle"))
print("MountHuddle")
kruskal.test(mean_prop ~ Trt2, data = filter(data_6, Behavior == "Mounted \nHuddle"))
```

```{r}
dunnTest(mean_prop ~ Trt2, data = filter(data_6, Behavior == "Side \nHuddle"), method = "holm")
```
```{r}
dunnTest(mean_prop ~ Trt2, data = filter(data_6, Behavior == "Activity"), method = "holm")
```
